♻️ Make using a Registry method
