+++ lamindb/_registry.py
@@ -1,14 +1,20 @@
 import builtins
 from typing import Iterable, List, NamedTuple, Optional, Union
+from uuid import UUID
 
+import dj_database_url
 import lamindb_setup as ln_setup
 import pandas as pd
 from django.core.exceptions import FieldDoesNotExist
+from django.db import connections
 from django.db.models import Manager, QuerySet
 from lamin_utils import logger
 from lamin_utils._lookup import Lookup
 from lamin_utils._search import search as base_search
+from lamindb_setup._init_instance import InstanceSettings
+from lamindb_setup._load_instance import get_owner_name_from_identifier
 from lamindb_setup.dev._docs import doc_args
+from lamindb_setup.dev._hub_core import load_instance
 from lnschema_core import Registry
 from lnschema_core.types import ListLike, StrField
 
@@ -334,6 +340,42 @@ def _queryset(cls: Union[Registry, QuerySet, Manager], **expressions) -> QuerySe
     return queryset
 
 
+def add_db_connection(isettings: InstanceSettings, using: str):
+    db_config = dj_database_url.config(
+        default=isettings.db, conn_max_age=600, conn_health_checks=True
+    )
+    db_config["TIME_ZONE"] = "UTC"
+    db_config["OPTIONS"] = {}
+    db_config["AUTOCOMMIT"] = True
+    connections.settings[using] = db_config
+
+
+@classmethod  # type: ignore
+@doc_args(Registry.using.__doc__)
+def using(
+    cls,
+    instance: str,
+) -> QuerySet:
+    owner, name = get_owner_name_from_identifier(instance)
+    load_result = load_instance(owner=owner, name=name)
+    if isinstance(load_result, str):
+        raise RuntimeError(
+            f"Fail to load instance {using}, please check your permission!"
+        )
+    instance_result, storage_result = load_result
+    isettings = InstanceSettings(
+        owner=owner,
+        name=name,
+        storage_root=storage_result["root"],
+        storage_region=storage_result["region"],
+        db=instance_result["db"],
+        schema=instance_result["schema_str"],
+        id=UUID(instance_result["id"]),
+    )
+    add_db_connection(isettings, using)
+    return QuerySet(model=cls, using=using)
+
+
 REGISTRY_UNIQUE_FIELD = {
     "storage": "root",
     "modality": "name",
@@ -435,6 +477,7 @@ METHOD_NAMES = [
     "lookup",
     "save",
     "from_values",
+    "using",
 ]
 
 if _TESTING:  # type: ignore
